+++ Sources/SwiftGenKit/Parsers/ColorsParser/ColorsCLRFileParser.swift
