+++ Sources/SwiftGenKit/Parsers/ColorsParser/ColorsJSONFileParser.swift
