+++ Sources/SwiftGenKit/Parsers/ColorsParser/ColorsTXTFileParser.swift
