+++ Sources/SwiftGenKit/Parsers/ColorsParser/ColorsXMLFileParser.swift
